LPS-77989 Move to internal
