+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper/src/main/java/com/liferay/portal/osgi/web/servlet/context/helper/internal/order/OrderBeforeAndAfterException.java
@@ -12,16 +12,16 @@
  * details.
  */
 
-package com.liferay.portal.osgi.web.servlet.context.helper.order;
+package com.liferay.portal.osgi.web.servlet.context.helper.internal.order;
 
 /**
  * @author Vernon Singleton
  * @author Juan Gonzalez
  */
-public class OrderMaxAttemptsException extends Exception {
+public class OrderBeforeAndAfterException extends Exception {
 
-	public OrderMaxAttemptsException(int limit) {
-		super("Exceeded the limit of " + limit + " maximum attempts");
+	public OrderBeforeAndAfterException(String configName, String name) {
+		super(configName + " cannot be both before and after " + name);
 	}
 
 }
