+++ modules/apps/static/portal-osgi-web/portal-osgi-web-servlet-context-helper/src/main/java/com/liferay/portal/osgi/web/servlet/context/helper/internal/order/OrderCircularDependencyException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.osgi.web.servlet.context.helper.order;
+package com.liferay.portal.osgi.web.servlet.context.helper.internal.order;
 
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.osgi.web.servlet.context.helper.definition.WebXMLDefinition;
